Move resource mapper to dedicated directory
