+++ src/core/services/resourceMapper/MockReferenceResolver.ts
