+++ src/views/dashboard/overview.tsx
@@ -4,7 +4,7 @@ import { FaFolderOpen } from 'react-icons/fa';
 import { ImmunizationStatusCard } from '../../components/dashboard/immunizationStatus/immunizationStatusCard';
 import { RecommendationCard } from '../../components/dashboard/recommendationCard/recommendationCard';
 import { ImmunizationRecommendation } from '../../core/models/ImmunizationRecommendation';
-import { useMapper } from '../../core/services/server/ResourceMapperContext';
+import { useMapper } from '../../core/services/resourceMapper/ResourceMapperContext';
 
 export function Overview() {
   const mapper = useMapper();
